requirements.txt
Django~=5.1.2
mysqlclient
aiohttp==3.11.12
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
streamlit==1.42.2
streamlit-jwt-authenticator==0.0.2
st-pages==1.0.1
altair~=5.5.0
pandas~=2.2.3
requests~=2.32.3